input: synaptics: s3508: Prioritize touches over button presses
Our touchscreen firmware commonly reports a button press and normal touch
simultaneously, causing both to be processed at the same time and bad UX
to ensue (due to an unintended button press taking over).

When there is a button press and touch detected at the same time, prioritize
the touch so that unintentional button presses are ignored. This reduces the
frequency that random button presses occur at when the user's finger quickly
moves from the display area to the key area (such as when scrolling through
long lists).

Signed-off-by: Sultanxda <sultanxda@gmail.com>